refactor: split account settings sections
